FIX: remove unneeded dependencies and fix tests.
