+++ droid-command-line/pom.xml
@@ -132,12 +132,7 @@
             <groupId>uk.gov.nationalarchives</groupId>
             <artifactId>droid-results</artifactId>
             <version>${project.version}</version>
-        </dependency>  	
-        <dependency>
-            <groupId>uk.gov.nationalarchives</groupId>
-            <artifactId>droid-core</artifactId>
-            <version>${project.version}</version>
-  	    </dependency>
+        </dependency>
         <dependency>
             <groupId>uk.gov.nationalarchives</groupId>
             <artifactId>droid-core-interfaces</artifactId>
@@ -153,11 +148,6 @@
             <artifactId>droid-report</artifactId>
             <version>${project.version}</version>
   	    </dependency>
-        <dependency>
-            <groupId>uk.gov.nationalarchives</groupId>
-            <artifactId>droid-container</artifactId>
-            <version>${project.version}</version>
-        </dependency>
         <dependency>
             <groupId>uk.gov.nationalarchives</groupId>
             <artifactId>droid-export-interfaces</artifactId>
@@ -191,46 +181,12 @@
             <artifactId>commons-io</artifactId>
         </dependency>
         <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-compress</artifactId>
-        </dependency>
-        <dependency>
             <groupId>commons-lang</groupId>
             <artifactId>commons-lang</artifactId>
         </dependency>
         <dependency>
             <groupId>joda-time</groupId>
             <artifactId>joda-time</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.jwat</groupId>
-            <artifactId>jwat-arc</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.jwat</groupId>
-            <artifactId>jwat-warc</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.jwat</groupId>
-            <artifactId>jwat-common</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.github.junrar</groupId>
-            <artifactId>junrar</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.github.stephenc.java-iso-tools</groupId>
-            <artifactId>loop-fs-iso-impl</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.ant</groupId>
-            <artifactId>ant-compress</artifactId>
-            <exclusions>
-                <exclusion>
-                    <groupId>com.sun</groupId>
-                    <artifactId>tools</artifactId>
-                </exclusion>
-            </exclusions>
         </dependency>
         <dependency>
             <groupId>javax.xml.bind</groupId>
@@ -259,11 +215,6 @@
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>de.waldheinz</groupId>
-            <artifactId>fat32-lib</artifactId>
-            <version>0.6.5</version>
         </dependency>
         <dependency>
             <groupId>org.hamcrest</groupId>
