+++ droid-command-line/src/test/java/uk/gov/nationalarchives/droid/command/action/FilterFieldCommandTest.java
@@ -46,8 +46,6 @@ import org.junit.Ignore;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
 
-import uk.gov.nationalarchives.droid.command.FilterFieldCommand;
-
 /**
  * @author rflitcroft
  *
